+++ src/fields_parse.rs
@@ -1,16 +1,15 @@
-extern crate quote;
-
-use quote::{quote, ToTokens};
-use syn::{FieldsNamed, FieldsUnnamed, Type};
+use quote::quote;
+use syn::{Fields, FieldsNamed, FieldsUnnamed, Type};
 
 #[derive(Debug)]
-pub enum ParserType {
+pub(crate) enum ParserType {
     Default,
     Split { separator: Option<String> },
     Custom(String),
 }
 
 impl ParserType {
+    // FIXME: use path for custom
     pub fn parse(data: &str) -> Self {
         match data {
             "default" => ParserType::Default,
@@ -20,16 +19,29 @@ impl ParserType {
     }
 }
 
-pub fn impl_parse_args_unnamed(
+pub(crate) fn impl_parse_args(
+    fields: &Fields,
+    self_variant: proc_macro2::TokenStream,
+    parser: &ParserType,
+) -> proc_macro2::TokenStream {
+    match fields {
+        Fields::Unit => self_variant,
+        Fields::Unnamed(fields) => {
+            impl_parse_args_unnamed(fields, self_variant, parser)
+        }
+        Fields::Named(named) => {
+            impl_parse_args_named(named, self_variant, parser)
+        }
+    }
+}
+
+pub(crate) fn impl_parse_args_unnamed(
     data: &FieldsUnnamed,
-    variant: impl ToTokens,
+    variant: proc_macro2::TokenStream,
     parser_type: &ParserType,
-) -> quote::__private::TokenStream {
-    let get_arguments = create_parser(
-        parser_type,
-        data.unnamed.iter().map(|f| &f.ty),
-        data.unnamed.len(),
-    );
+) -> proc_macro2::TokenStream {
+    let get_arguments =
+        create_parser(parser_type, data.unnamed.iter().map(|f| &f.ty));
     let iter = (0..data.unnamed.len()).map(syn::Index::from);
     let mut initialization = quote! {};
     for i in iter {
@@ -44,17 +56,14 @@ pub fn impl_parse_args_unnamed(
     res
 }
 
-pub fn impl_parse_args_named(
+pub(crate) fn impl_parse_args_named(
     data: &FieldsNamed,
-    variant: impl ToTokens,
+    variant: proc_macro2::TokenStream,
     parser_type: &ParserType,
-) -> quote::__private::TokenStream {
-    let get_arguments = create_parser(
-        parser_type,
-        data.named.iter().map(|f| &f.ty),
-        data.named.len(),
-    );
-    let i = (0..data.named.len()).map(syn::Index::from);
+) -> proc_macro2::TokenStream {
+    let get_arguments =
+        create_parser(parser_type, data.named.iter().map(|f| &f.ty));
+    let i = (0..).map(syn::Index::from);
     let name = data.named.iter().map(|f| f.ident.as_ref().unwrap());
     let res = quote! {
         {
@@ -67,26 +76,30 @@ pub fn impl_parse_args_named(
 
 fn create_parser<'a>(
     parser_type: &ParserType,
-    mut types: impl Iterator<Item = &'a Type>,
-    count_args: usize,
-) -> quote::__private::TokenStream {
+    mut types: impl ExactSizeIterator<Item = &'a Type>,
+) -> proc_macro2::TokenStream {
     let function_to_parse = match parser_type {
-        ParserType::Default => match count_args {
+        ParserType::Default => match types.len() {
             1 => {
-                let ty = types.next().expect("count_args != types.len()");
-                quote! { (|s: String| {
-                    let res = <#ty>::from_str(&s)
-                        .map_err(|e|ParseError::IncorrectFormat({ let e: Box<dyn std::error::Error + Send + Sync + 'static> = e.into(); e }))?;
-                    Ok((res, ))
-                 })
+                let ty = types.next().unwrap();
+                quote! {
+                    (
+                        |s: String| {
+                            let res = <#ty>::from_str(&s)
+                                .map_err(|e| ParseError::IncorrectFormat(e.into()))?;
+
+                            Ok((res,))
+                        }
+                    )
                 }
             }
-            _ => quote! { compile_error!("Expected exactly 1 argument") },
+            _ => {
+                quote! { compile_error!("Default parser works only with exactly 1 field") }
+            }
         },
         ParserType::Split { separator } => parser_with_separator(
             &separator.clone().unwrap_or_else(|| " ".to_owned()),
             types,
-            count_args,
         ),
         ParserType::Custom(s) => {
             let path = syn::parse_str::<syn::Path>(s).unwrap_or_else(|_| {
@@ -95,6 +108,7 @@ fn create_parser<'a>(
             quote! { #path }
         }
     };
+
     quote! {
         let arguments = #function_to_parse(args)?;
     }
@@ -102,31 +116,46 @@ fn create_parser<'a>(
 
 fn parser_with_separator<'a>(
     separator: &str,
-    types: impl Iterator<Item = &'a Type>,
-    count_args: usize,
-) -> quote::__private::TokenStream {
-    let inner = quote! { let mut splited = s.split(#separator); };
-    let i = 0..count_args;
-    let inner2 = quote! {
-        #(<#types>::from_str(splited.next().ok_or(ParseError::TooFewArguments {
-            expected: #count_args,
-            found: #i,
-            message: format!("Expected but not found arg number {}", #i + 1),
-        })?).map_err(|e|ParseError::IncorrectFormat({ let e: Box<dyn std::error::Error + Send + Sync + 'static> = e.into(); e }))?,)*
+    types: impl ExactSizeIterator<Item = &'a Type>,
+) -> proc_macro2::TokenStream {
+    let expected = types.len();
+    let res = {
+        let found = 0usize..;
+        quote! {
+            (
+                #(
+                    {
+                        let s = splitted.next().ok_or(ParseError::TooFewArguments {
+                            expected: #expected,
+                            found: #found,
+                            message: format!("Expected but not found arg number {}", #found + 1),
+                        })?;
+
+                        <#types>::from_str(s).map_err(|e| ParseError::IncorrectFormat(e.into()))?
+                    }
+                ),*
+            )
+        }
     };
+
     let res = quote! {
-        (|s: String| {
-            #inner
-            let res = (#inner2);
-            match splited.next() {
-                Some(d) => Err(ParseError::TooManyArguments {
-                    expected: #count_args,
-                    found: #count_args + 1,
-                    message: format!("Excess argument: {}", d),
-                }),
-                None => Ok(res)
+        (
+            |s: String| {
+                let mut splitted = s.split(#separator);
+
+                let res = #res;
+
+                match splitted.next() {
+                    Some(d) => Err(ParseError::TooManyArguments {
+                        expected: #expected,
+                        found: #expected + 1,
+                        message: format!("Excess argument: {}", d),
+                    }),
+                    None => Ok(res)
+                }
             }
-        })
+        )
     };
+
     res
 }