+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Removed
+
+- Remove `derive(DialogueState)` macro
+
+### Changed
+
+- `#[command(rename = "a_name_that_is_not_a_case_name")]` doesn't work anymore
+
 ## 0.6.3 - 2022-07-19
 
 ### Fixed